video-ai-filter/pyproject.toml
[project]
name = "video-ai-filter"
version = "0.1.0"
description = "REST API: sample video frames and moderate via OpenRouter Qwen3 VL"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.12",
    "opencv-python-headless>=4.10.0",
    "openai>=1.55.0",
    "pydantic-settings>=2.6.0",
    "faster-whisper>=1.0.0",
]
